+++ packages/flutter_robusta/lib/src/router.dart
@@ -4,6 +4,7 @@ import 'package:copy_with_extension/copy_with_extension.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router_plus/go_router_plus.dart';
+import 'package:meta/meta.dart';
 
 part 'router.g.dart';
 
@@ -25,9 +26,10 @@ typedef RedirectorFactory = _Factory<Redirector>;
 typedef NavigatorObserverFactory = _Factory<NavigatorObserver>;
 
 /// {@template router.settings}
-/// Instances of this class support to settings router
+/// Instances of this class supports router settings.
 /// {@endtemplate}
 @CopyWith(skipFields: true)
+@sealed
 class RouterSettings {
   /// {@macro router.settings}
   RouterSettings({