+++ packages/robusta_runner/lib/src/runner/exception.dart
@@ -5,11 +5,11 @@ part of '../runner.dart';
 class RunnerException implements Exception {
   RunnerException._(this._msg);
 
-  /// Throw in case user add duplicate extensions.
+  /// Throw in case user added duplicate extensions.
   factory RunnerException.duplicateExtension(Type e) =>
       RunnerException._('Duplicate extension $e');
 
-  /// Throw in case user get extension but it not exist.
+  /// Throw in case user gets extension but it not exist.
   factory RunnerException.extensionNotExist(Type e) => RunnerException._(
         'Extension $e not exist, do you forget to add it to runner extensions?',
       );