+++ packages/robusta_runner/test/src/runner_test.dart
@@ -147,7 +147,7 @@ void main() {
       expect(
         () => Runner(
           extensions: [
-            TestExtension(),
+            ImplementingCallbackExtension(),
             TestDependenceExtension(),
           ],
         ),